Attempt to use PRAGMA busy_timeout
Ideally, fixes #4247. Instead of immediately failing on the database
being busy (via usage by another process), Stack will now have a 2
second pause and then try again. The vast majority of interactions with
the database are much faster than 2 seconds. The only exception is the
population of the Hackage index cache, but that's hopefully rare enough,
and touching few enough tables, to not be a problem.