Use --force-local option for tar
Without the above option, the tar commands results in this error in an
Windows environment:

/usr/bin/tar: Cannot connect to D: resolve failed

That's because when there is a colon (:) in a filename, tar treats as
though the file is in a different machine. And this comes into act for
windows paths like `C://`.

The --force-local option informs GNU Tar to treat the archive file as
local even if the file path has colon in it.