Don't compare generated cabal files
For an example of this bug, see
https://github.com/commercialhaskell/stack/issues/5045.

Problem: pantry was including the hash of generated cabal files in lock
files, and then when newly generated cabal files (based on new hpack
versions) mismatched, it considered it a different package. Solution: do
not store the hash of the generated cabal file, but instead of the
source hpack file.