Clear (another) compiler warning
Clears:

~~~
src\Pantry\Types.hs:567:9: warning: [-Wincomplete-uni-patterns]
    Pattern match(es) are non-exhaustive
    In a lambda abstraction:
        Patterns of type ‘[(Repo, RawPackageMetadata)]’ not matched: []
    |
567 |   fmap (\xs@((repo, _):_) -> AggregateRepo (rToSimpleRepo repo) (fmap (first repoSubdir) xs))
    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
~~~

I assume that the lambda abstraction should not be a partial function. I use `Maybe` to make it complete, and then replace the `fmap` with a `mapMaybe`.

I also found the original `toAggregateRepos` function, with its two lambdas, a little opaque. So, I also replaced the second lambda with a named function.