Separate true `HackageSecurityConfig` from (renamed) `PackageIndexConfig`
This pull request proposes a relatively minor, but breaking, change to Pantry's API. There is a corresponding pull request for Stack that adopts this version of the Pantry API.

The motivation for this proposal is the request at Stack issue [#5870](https://github.com/commercialhaskell/stack/issues/5870).

What is currently referred to as `HackageSecurityConfig` is better named `PackageIndexConfig`, comprising a true `HackageSecuityConfig` value and a download prefix (`picDownloadPrefix :: !Text`).

What is currently referred to as `defaultHackageSecurityConfig` is renamed `defaultPackageIndexConfig`.

True `HackageSecurityConfig` (strictly, `WithJSONWarnings HackageSecurityConfig`) has its own instance of `FromJSON` and its own default: `defaultHackageSecurityConfig`.

Meeting the ask in the Stack issue, the instance of `FromJSON` for `PackageIndexConfig` now assigns a default value of `defaultHackageSecurityConfig` if the `hackage-security` key is absent from the JSON object.

The field of `PackageConfig` named `pcHackageSecuity :: !HackageSecurityConfig` is renamed `pcPackageIndex :: !PackageIndexConfig`.

For completeness, `defaultDownloadPrefix` is also exposed.