Fix #69 Don't hard code "stack" as Hpack program name
Like Stack, also refers to Hpack as 'Hpack' (unless referring to the command 'hpack') and refers to Cabal files as 'Cabal files' rather than '.cabal' files (unless the context is the extension to the files).

`Pantry.Storage.findOrGenerateCabalFile` is a similar function to `Pantry.findOrGenerateCabalFile` but (1) the module is not exposed by the library and (2) the other modules in the library do not use `Pantry.Storage.findOrGenerateCabalFile` - proven by hiding it on import.