Extend Hpack.AlreadyGeneratedByNewerHpack warning
Fixes https://github.com/commercialhaskell/stack/issues/6057.

Hpack does not over-write an existing Cabal file, if it was generated by a more recent version of Hpack. Pantry, when it is using its built-in version of Hpack, warns when this occurs but does not consider it fatal - it will go on to use the existing Cabal file that was not over-written.

This extends the warning for `Hpack.AlreadyGeneratedByNewerHpack` to include the existing advice given for `Hpack.ExistingCabalFileWasModifiedManually`, which is equally applicable.